Trajectory Highlighter

InteractiveChaos.trajectory_highlighter — Function

trajectory_highlighter(datasets, vals; kwargs...)

Open an interactive application for highlighting specific datasets and properties of these datasets. datasets is a vector of anything from DynamicalSystems.jl that supports plot_dataset (currently Dataset or Matrix). Each dataset corresponds to a specific value from vals (a Vector{<:Real}). The value of vals gives each dataset a specific color based on a colormap.

The application is composed of two scenes: the left scene plots the datasets, while the right scene plots the histogram of the vals. The function returns the two scenes data_scene, hist_scene.

Interaction

Clicking on a bin of the histogram plot will "highlight" all data whose value belongs in that bin. Here highlighting actually means "hidding" (i.e. reducing their alpha value) all other data besides the ones you want to highlight. Clicking on empty space on the histogram plot will reset highlighting.

Clicking on a plotted series in the left window will highlight this series as well as the histogram bin that contains its value. Clicking on empty space will reset the highlighting.

Keyword Arguments

nbins = 10, closed = :left : used in producing the histogram.

α = 0.05 : the alpha value of the hidden data.

hα = 0.2 : the alpha value of the hidden histogram bins.

cmap = :viridis : the colormap used.

hname = "value" : name for the histogram axis.

kwargs... : Anything else is propagated to plot_dataset.

2D Version:

3D Version: