Interactive trajectory evolution

InteractiveChaos.interactive_evolution — Function

interactive_evolution(ds::DynamicalSystem, u0s; kwargs...)

Launch an interactive application that can evolve the initial conditions u0s (vector of vectors) of the given dynamical system. All initial conditions are evolved in parallel and at exactly the same time. Two controls allow you to pause/resume the evolution and to adjust the speed. The application can run forever (trajectories are computed on demand).

The function returns scene, main, layout, obs. scene is the overarching scene (the entire GUI) and can be recorded. main is the actual plot of the trajectory, that allows adding additional plot elements or controlling labels, ticks, etc. layout is the overarching layout, that can be used to add additional plot panels. obs is a vector of observables, each containing the current state of the trajectory.

Keywords

idxs = 1:min(dimension(ds), 3) : Which variables to plot (up to three can be chosen).

colors : The color for each trajectory. Random colors are chosen by default.

lims : A tuple of tuples (min, max) for the axis limits. If not given, they are automatically deduced by evolving each of u0s 1000 units and picking most extreme values (limits cannot be adjust after application is launched).

plotkwargs = NamedTuple() : A named tuple of keyword arguments propagated to the plotting function (lines for continuous, scatter for discrete systems).

tail = 1000 : Length of plotted trajectory (in step units).

diffeq = DynamicalSystems.CDS_KWARGS : Named tuple of keyword arguments propagated to the solvers of DifferentialEquations.jl (for continuous systems). Because trajectories are not pre-computed and interpolated, it is recommended to use a combination of arguments that limit maximum stepsize, to ensure smooth curves. For example:
using OrdinaryDiffEq
diffeq = (alg = Tsit5(), dtmax = 0.01)

To generate the above video run

using InteractiveChaos
using DynamicalSystems, Makie
using OrdinaryDiffEq
ds = Systems.henonheiles()

u0s = [[0.0, -0.25, 0.42081, 0.0],
[0.0, 0.1, 0.5, 0.0],
[0.0, -0.31596, 0.354461, 0.0591255]]

diffeq = (alg = Vern9(), dtmax = 0.01)
idxs = (1, 2, 4)
colors = ["#233B43", "#499cbf", "#E84646"]

scene, main, obs = interactive_evolution(
    ds, u0s; idxs, tail = 10000, diffeq, colors
)